Atualizando script de criação do banco de dados.

+++ assets/database/database_creation_script.sql
@@ -48,8 +48,8 @@ CREATE TABLE `payment` (
   `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `rental_id` bigint(20) NOT NULL,
   `rent_value` decimal(15,2) NOT NULL,
-  `competence_month` tinyint(2) unsigned NOT NULL,
-  `competence_year` smallint(4) unsigned NOT NULL,
+  `competence_month` int(2) unsigned NOT NULL,
+  `competence_year` int(4) unsigned NOT NULL,
   `status` enum('allegedly_paid','confirmed','rejected') NOT NULL DEFAULT 'allegedly_paid',
   `created_at` timestamp NOT NULL DEFAULT current_timestamp(),
   PRIMARY KEY (`id`),
@@ -264,4 +264,4 @@ CREATE TABLE `visit` (
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2023-11-19 18:03:47
+-- Dump completed on 2023-11-19 18:09:07
